Moved validator logic to core file
